+++ src/runtime/parallel/slave.js
@@ -3,7 +3,6 @@ import { formatLocation } from '../../formatter/helpers'
 import commandTypes from './command_types'
 import EventEmitter from 'events'
 import Promise from 'bluebird'
-import readline from 'readline'
 import serializeError from 'serialize-error'
 import StackTraceFilter from '../stack_trace_filter'
 import supportCodeLibraryBuilder from '../../support_code_library_builder'
@@ -20,30 +19,29 @@ const EVENTS = [
   'test-case-finished',
 ]
 
-function replacerSerializeErrors(key, value) {
-  if (_.isError(value)) {
-    return serializeError(value)
+function serializeResultExceptionIfNecessary(data) {
+  if (
+    data.result &&
+    data.result.exception &&
+    _.isError(data.result.exception)
+  ) {
+    data.result.exception = serializeError(data.result.exception)
   }
-  return value
 }
 
 export default class Slave {
-  constructor({ cwd, stdin, stdout }) {
+  constructor({ cwd, exit, sendMessage }) {
     this.initialized = false
-    this.stdin = stdin
-    this.stdout = stdout
     this.cwd = cwd
+    this.exit = exit
+    this.sendMessage = sendMessage
     this.eventBroadcaster = new EventEmitter()
     this.stackTraceFilter = new StackTraceFilter()
     EVENTS.forEach(name => {
-      this.eventBroadcaster.on(name, data =>
-        this.stdout.write(
-          JSON.stringify(
-            { command: commandTypes.EVENT, name, data },
-            replacerSerializeErrors
-          ) + '\n'
-        )
-      )
+      this.eventBroadcaster.on(name, data => {
+        serializeResultExceptionIfNecessary(data)
+        this.sendMessage({ command: commandTypes.EVENT, name, data })
+      })
     })
   }
 
@@ -63,7 +61,7 @@ export default class Slave {
       this.stackTraceFilter.filter()
     }
     await this.runTestRunHooks('beforeTestRunHookDefinitions', 'a BeforeAll')
-    this.stdout.write(JSON.stringify({ command: commandTypes.READY }) + '\n')
+    this.sendMessage({ command: commandTypes.READY })
   }
 
   async finalize() {
@@ -71,25 +69,17 @@ export default class Slave {
     if (this.filterStacktraces) {
       this.stackTraceFilter.unfilter()
     }
-    process.exit()
+    this.exit()
   }
 
-  parseMasterLine(line) {
-    const input = JSON.parse(line)
-    if (input.command === 'initialize') {
-      this.initialize(input)
-    } else if (input.command === 'finalize') {
+  receiveMessage(message) {
+    if (message.command === 'initialize') {
+      this.initialize(message)
+    } else if (message.command === 'finalize') {
       this.finalize()
-    } else if (input.command === 'run') {
-      this.runTestCase(input)
+    } else if (message.command === 'run') {
+      this.runTestCase(message)
     }
-  }
-
-  async run() {
-    this.rl = readline.createInterface({ input: this.stdin })
-    this.rl.on('line', line => {
-      this.parseMasterLine(line)
-    })
   }
 
   async runTestCase({ testCase, retry, skip }) {
@@ -102,7 +92,7 @@ export default class Slave {
       worldParameters: this.worldParameters,
     })
     await testCaseRunner.run()
-    this.stdout.write(JSON.stringify({ command: commandTypes.READY }) + '\n')
+    this.sendMessage({ command: commandTypes.READY })
   }
 
   async runTestRunHooks(key, name) {