+++ src/runtime/test_case_runner.js
@@ -127,6 +127,10 @@ export default class TestCaseRunner {
     return this.result.status !== Status.PASSED
   }
 
+  shouldSkipHook(isBeforeHook) {
+    return this.skip || (this.isSkippingSteps() && isBeforeHook)
+  }
+
   shouldUpdateStatus(testStepResult) {
     switch (testStepResult.status) {
       case Status.FAILED:
@@ -166,16 +170,24 @@ export default class TestCaseRunner {
     this.emitPrepared()
     this.emit('test-case-started', {})
     for (let i = 0; i <= this.retry; i++) {
-      await this.runHooks(this.beforeHookDefinitions, {
-        sourceLocation: this.testCaseSourceLocation,
-        pickle: this.testCase.pickle,
-      })
+      await this.runHooks(
+        this.beforeHookDefinitions,
+        {
+          sourceLocation: this.testCaseSourceLocation,
+          pickle: this.testCase.pickle,
+        },
+        true
+      )
       await this.runSteps()
-      await this.runHooks(this.afterHookDefinitions, {
-        sourceLocation: this.testCaseSourceLocation,
-        pickle: this.testCase.pickle,
-        result: this.result,
-      })
+      await this.runHooks(
+        this.afterHookDefinitions,
+        {
+          sourceLocation: this.testCaseSourceLocation,
+          pickle: this.testCase.pickle,
+          result: this.result,
+        },
+        false
+      )
       if (this.result.status !== Status.FAILED) {
         break
       }
@@ -189,17 +201,17 @@ export default class TestCaseRunner {
     return this.result
   }
 
-  async runHook(hookDefinition, hookParameter) {
-    if (this.skip) {
+  async runHook(hookDefinition, hookParameter, isBeforeHook) {
+    if (this.shouldSkipHook(isBeforeHook)) {
       return { status: Status.SKIPPED }
     }
     return this.invokeStep(null, hookDefinition, hookParameter)
   }
 
-  async runHooks(hookDefinitions, hookParameter) {
+  async runHooks(hookDefinitions, hookParameter, isBeforeHook) {
     await Promise.each(hookDefinitions, async hookDefinition => {
       await this.aroundTestStep(() =>
-        this.runHook(hookDefinition, hookParameter)
+        this.runHook(hookDefinition, hookParameter, isBeforeHook)
       )
     })
   }